Introduce our own responsive button
