+++ packages/react-admin/src/mui/button/ListButton.js
@@ -1,60 +1,19 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import compose from 'recompose/compose';
-import Button from 'material-ui/Button';
 import ActionList from 'material-ui-icons/List';
-import { withStyles } from 'material-ui/styles';
-import classnames from 'classnames';
 
 import Link from '../Link';
-import translate from '../../i18n/translate';
-import Responsive from '../layout/Responsive';
+import Button from './Button';
 
-const styles = {
-    link: {
-        display: 'inline-flex',
-        alignItems: 'center',
-    },
-    label: {
-        marginLeft: '0.5em',
-    },
-};
-
-const ListButton = ({
-    basePath = '',
-    className,
-    classes = {},
-    label = 'ra.action.list',
-    translate,
-    ...rest
-}) => (
-    <Button
-        className={classnames(classes.link, className)}
-        component={Link}
-        color="primary"
-        to={basePath}
-        {...rest}
-    >
+const ListButton = ({ basePath = '', label = 'ra.action.list', ...rest }) => (
+    <Button component={Link} to={basePath} label={label} {...rest}>
         <ActionList />
-        <Responsive
-            small={<span />}
-            medium={
-                <span className={classes.label}>
-                    {label && translate(label)}
-                </span>
-            }
-        />
     </Button>
 );
 
 ListButton.propTypes = {
     basePath: PropTypes.string,
-    className: PropTypes.string,
-    classes: PropTypes.object,
     label: PropTypes.string,
-    translate: PropTypes.func.isRequired,
 };
 
-const enhance = compose(withStyles(styles), translate);
-
-export default enhance(ListButton);
+export default ListButton;
